backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
playwright
scikit-learn
numpy
openai
google-generativeai
